- Refactor learning rate schedule code to live in `keras/optimizers/schedules` (to match the API namespace). - Export optimizers to `keras.optimizers.__init__` to resolve discrepancy between tf.keras and keras packages. - Export schedules to `keras.optimizers.schedules.__init__` to resolve discrepancy between tf.keras and keras packages.
PiperOrigin-RevId: 435722952
