+++ keras/api/BUILD
@@ -89,7 +89,6 @@ keras_packages = [
     "keras.mixed_precision.loss_scale_optimizer",
     "keras.mixed_precision.policy",
     "keras.models",
-    "keras.optimizers.learning_rate_schedule",
     "keras.optimizers.legacy.adadelta",
     "keras.optimizers.legacy.adagrad",
     "keras.optimizers.legacy.adam",
@@ -117,6 +116,7 @@ keras_packages = [
     "keras.optimizers.optimizer_v2.nadam",
     "keras.optimizers.optimizer_v2.optimizer_v2",
     "keras.optimizers.optimizer_v2.rmsprop",
+    "keras.optimizers.schedules.learning_rate_schedule",
     "keras.optimizers",
     "keras.premade_models.linear",
     "keras.premade_models.wide_deep",
