+++ keras/api/golden/v1/tensorflow.keras.experimental.-cosine-decay.pbtxt
@@ -1,7 +1,7 @@
 path: "tensorflow.keras.experimental.CosineDecay"
 tf_class {
-  is_instance: "<class \'keras.optimizers.learning_rate_schedule.CosineDecay\'>"
-  is_instance: "<class \'keras.optimizers.learning_rate_schedule.LearningRateSchedule\'>"
+  is_instance: "<class \'keras.optimizers.schedules.learning_rate_schedule.CosineDecay\'>"
+  is_instance: "<class \'keras.optimizers.schedules.learning_rate_schedule.LearningRateSchedule\'>"
   is_instance: "<type \'object\'>"
   member_method {
     name: "__init__"
