+++ keras/api/golden/v2/tensorflow.keras.optimizers.schedules.-cosine-decay.pbtxt
@@ -1,7 +1,7 @@
 path: "tensorflow.keras.optimizers.schedules.CosineDecay"
 tf_class {
-  is_instance: "<class \'keras.optimizers.learning_rate_schedule.CosineDecay\'>"
-  is_instance: "<class \'keras.optimizers.learning_rate_schedule.LearningRateSchedule\'>"
+  is_instance: "<class \'keras.optimizers.schedules.learning_rate_schedule.CosineDecay\'>"
+  is_instance: "<class \'keras.optimizers.schedules.learning_rate_schedule.LearningRateSchedule\'>"
   is_instance: "<type \'object\'>"
   member_method {
     name: "__init__"
