+++ keras/dtensor/optimizers.py
@@ -15,13 +15,13 @@
 """DTensor specific Keras optimizers."""
 
 from keras.dtensor import dtensor_api as dtensor
-from keras.optimizers import learning_rate_schedule
 from keras.optimizers.optimizer_experimental import adadelta
 from keras.optimizers.optimizer_experimental import adagrad
 from keras.optimizers.optimizer_experimental import adam
 from keras.optimizers.optimizer_experimental import optimizer as optimizer_lib
 from keras.optimizers.optimizer_experimental import rmsprop
 from keras.optimizers.optimizer_experimental import sgd
+from keras.optimizers.schedules import learning_rate_schedule
 
 import tensorflow.compat.v2 as tf
