+++ keras/optimizers/legacy_learning_rate_decay.py
@@ -17,7 +17,7 @@
 import tensorflow.compat.v2 as tf
 
 import functools
-from keras.optimizers import learning_rate_schedule
+from keras.optimizers.schedules import learning_rate_schedule
 from tensorflow.python.util.tf_export import tf_export
