+++ keras/optimizers/optimizer_experimental/BUILD
@@ -27,7 +27,7 @@ py_library(
     srcs_version = "PY3",
     deps = [
         "//:expect_tensorflow_installed",
-        "//keras/optimizers:learning_rate_schedule",
+        "//keras/optimizers/schedules:learning_rate_schedule",
     ],
 )
