+++ keras/optimizers/optimizer_experimental/optimizer_test.py
@@ -9,7 +9,6 @@ import re
 from absl import logging
 from absl.testing import parameterized
 import keras
-from keras.optimizers import learning_rate_schedule
 from keras.optimizers.optimizer_experimental import adadelta as adadelta_new
 from keras.optimizers.optimizer_experimental import adagrad as adagrad_new
 from keras.optimizers.optimizer_experimental import adam as adam_new
@@ -25,6 +24,7 @@ from keras.optimizers.optimizer_v2 import adam as adam_old
 from keras.optimizers.optimizer_v2 import ftrl as ftrl_old
 from keras.optimizers.optimizer_v2 import gradient_descent as sgd_old
 from keras.optimizers.optimizer_v2 import rmsprop as rmsprop_old
+from keras.optimizers.schedules import learning_rate_schedule
 from keras.utils import losses_utils
 import numpy as np
 import tensorflow.compat.v2 as tf
