+++ keras/optimizers/optimizer_v2/adam_test.py
@@ -21,7 +21,7 @@ import numpy as np
 from keras.testing_infra import test_combinations
 from keras.optimizers import optimizer_v1
 from keras.optimizers.optimizer_v2 import adam
-from keras.optimizers import learning_rate_schedule
+from keras.optimizers.schedules import learning_rate_schedule
 
 
 def adam_update_numpy(param,
