+++ keras/optimizers/schedules/__init__.py
@@ -12,13 +12,9 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 # ==============================================================================
-"""Legacy optimizer package."""
+"""Learning rate schedule API."""
 
-from keras.optimizers.legacy.adadelta import Adadelta
-from keras.optimizers.legacy.adagrad import Adagrad
-from keras.optimizers.legacy.adam import Adam
-from keras.optimizers.legacy.adamax import Adamax
-from keras.optimizers.legacy.ftrl import Ftrl
-from keras.optimizers.legacy.nadam import Nadam
-from keras.optimizers.legacy.rmsprop import RMSprop
-from keras.optimizers.legacy.sgd import SGD
+from keras.optimizers.schedules.learning_rate_schedules import ExponentialDecay
+from keras.optimizers.schedules.learning_rate_schedules import InverseTimeDecay
+from keras.optimizers.schedules.learning_rate_schedules import PiecewiseConstantDecay
+from keras.optimizers.schedules.learning_rate_schedules import PolynomialDecay
